Refactor managed aggregator oracle tests
